Framework: Move withEditorSettings file to its own directory
